Add some failing unit tests for rectifying scope. Explanation follows.
In order to support ng-template, the existing implementation would require creating a separate ts file for each inner template. Each file would not be able to import an existing interface to define the scope, so each scope variable would need to be manually declared. This means the scope type checking would not be linked to the scope implementation, so it provides little value. This limitation also exists for the current template type checking.

I'll need to refactor this program so that it works on interfaces, and the generated code will declare scope variables implementing the interfaces. This requires re-writing AngularJS expressions, so that scope-level identifiers become TS member expressions, e.g. "someScopeValue" becomes "__scope_1.someScopeValue".

Other changes in this commit:
- Add core.logObject()
- Expand smorgasboard example a little
- Remove some commented code