Replace "jade" with "pug".
Unfortunately it seems that "include" statements will expect the included file to have an extension of ".pug", unless you explicitly write it like "include ./card.jade". Even then, it will emit a warning to stdout. This is not ideal for legacy projects, so maybe I'll actually treat .jade and .pug as separate template types.