Fix #10: an element that creates multiple "scopes", e.g. one containg both "ng-if" and "ng-controller" directives, would only break out of the last scope, instead of all scopes created at the same level. This lead to nodes being added as siblings of the inner scope, instead of the outer scope.
This is fixed by tracking the element used when creating each scope, and when a new scope is created, checking if it is for the same element as the previous scope. If so, we do a sneaky "pop()" to get rid of the previous scope, and only retain the most recent scope on the scope stack.